Compilation errors are fixed. Gives runtime error in KATTIS. Error is 100% in Player::play() method.

+++ TTT3D/player.hpp
@@ -21,7 +21,7 @@ public:
     uint8_t min_p;
 
     GameState play(const GameState &pState, const Deadline &pDue);
-    int alphabeta(const GameState &pState, uint8_t player, int depth, double alpha, double beta);
+    int alphabeta(const GameState &pState, uint8_t player, int depth, int alpha, int beta);
     int evaluation(const GameState &state);
 };
